Fixes from PR conversion
